airflow_docker/requirements.txt
apache-airflow==2.10.5
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
frozendict==2.4.6
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pip==25.0.1
platformdirs==4.3.7
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.3.0
requests==2.32.3
setuptools==78.1.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
yfinance==0.2.55